requirements.txt
flask==2.2.2
python-dotenv==0.21.0
psycopg2-binary
flask_cors==3.0.10